Separate Bottom Bar for StaticQuery
